feat(ManageProduct) ♻ Refactor UI layer
